updated clima app to show more info and change background image at each 20s using timer/uiview animations

+++ Swift 4 Course/Apps/Clima-App/Clima/Models/WindData.swift
@@ -10,9 +10,9 @@ import Foundation
 
 class WindData {
     let degrees: Int
-    let speed: Float
+    let speed: Int
     
-    init(degrees: Int, speed: Float) {
+    init(degrees: Int, speed: Int) {
         self.degrees = degrees
         self.speed = speed
     }
